Authors: itamarst, therve
Reviewer: exarkun
Fixes #2341

Create the base class twisted.internet.process._BaseProcess, used to merge most
of the code duplicated between Process and PTYProcess. Those classes now
inherit from this one. This helped to fix a few bugs, in particular one with
setuid management in PTYProcess.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@21081 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb

+++ twisted/internet/_dumbwin32proc.py
@@ -243,6 +243,8 @@ class Process(_pollingfile._PollingTimer):
 
 
     def signalProcess(self, signalID):
+        if self.closed:
+            raise error.ProcessExitedAlready()
         if signalID in ("INT", "TERM", "KILL"):
             win32process.TerminateProcess(self.hProcess, 1)
